Implement iterator-wise data gathering

+++ src/data_preprocessing/string_pattern_searching.py
@@ -1,4 +1,5 @@
 import re
+import numpy as np
 
 import pandas as pd
 
@@ -7,9 +8,9 @@ class StringPatternSearching:
     @classmethod
     def extract_numbers(cls, text: str) -> pd.Series:
         pairs = re.findall(r'(\w+?)(\d+)', text)
-        numbers = [int(value) for _, value in pairs]
+        numbers = [float(value) for _, value in pairs]
 
-        return pd.Series(numbers)
+        return pd.Series(numbers, dtype=np.float16)
 
     @classmethod
     def split_by_camelcase(cls, text: str) -> pd.Series:
